Move player standings solver
