+++ Slask.Domain/Groups/GroupUtility/AdvancingPlayersSolver.cs
@@ -1,6 +1,7 @@
 ﻿using Slask.Domain.Groups.Bases;
 using Slask.Domain.Rounds.Bases;
 using Slask.Domain.Utilities;
+using Slask.Utilities.PlayerStandingsSolver;
 using System;
 using System.Collections.Generic;
