+++ Slask.Domain/Rounds/RoundRobinRound.cs
@@ -1,6 +1,7 @@
-﻿using Slask.Domain.Groups;
+using Slask.Domain.Groups;
 using Slask.Domain.Groups.Bases;
 using Slask.Domain.Rounds.Bases;
+using Slask.Utilities.PlayerStandingsSolver;
 using System;
 using System.Linq;
