+++ Slask.SpecFlow.IntegrationTests/DomainTests/UtilityTests/PlayerStandingsSolverSteps.cs
@@ -3,6 +3,7 @@ using Slask.Common;
 using Slask.Domain.Groups.Bases;
 using Slask.Domain.Groups.GroupUtility;
 using Slask.SpecFlow.IntegrationTests.DomainTests.GroupTests;
+using Slask.Utilities.PlayerStandingsSolver;
 using System;
 using System.Collections.Generic;
 using TechTalk.SpecFlow;
